+++ docs/changes.md
@@ -21,6 +21,8 @@ Check example `basic/mesh_custom.py`
 ### `addons.py`
 - added `addSplineTool()` to interactively spline points in space
 - added `labelRotation` in addScalarBar3D
+- added `xShiftAlongY` keywords in `Axes` to slide the whole axis position along another axis
+- `addScalarBar3D()` can now render categorical data
 
 ---
 ### `colors.py`
@@ -38,6 +40,7 @@ Check example `basic/mesh_custom.py`
 - added `computeWorldPosition(point2d)` to get the 3d point in the scene from a screen 2d point
 - added `addSplineTool()` to interactively spline points in space
 - small fix in `plotter.add()` for offscreen mode.
+- added `topicture()` to render a scene into a `Picture` object (to crop, mirror etc)
 
 ---
 ### `picture.py`
@@ -46,6 +49,7 @@ Check example `basic/mesh_custom.py`
 
 ---
 ### `pointcloud.py`
+- added `cutWithBox()`, `cutWithLine()`, `cutWithSphere()` and `cutWithCylinder()` methods
 
 ---
 ### `pyplot.py`